test_bot/pyproject.toml
[tool.poetry]
name = "test-bot"
version = "0.0.0"
description = "A test bot for mafic."
authors = ["ooliver1 <oliverwilkes2006@icloud.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
nextcord = {extras = ["speed"], version = "^2.2.0"}
ooliver-botbase = "^1.22.3"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
isort = "^5.10.1"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
